Staging (#27)
* dynamic_port_selection_work

* undefined_string_issue

* baud_rate_selection_part_added

* serial_terminal_scroll_feature

* serial_data_input_show_with_different_color

* serial management in separate file

* Refactor serial_handler.py to extract and print
untrained messages.
