+++ serial_communication/web/app.py
@@ -1,16 +1,15 @@
 
-#$ last work 02/Nov/23 [12:41 AM]
-## version 1.0.4
-## Release Note : CW sending public IP
+#$ last work 05/Nov/23 [01:41 PM]
+## version 1.0.5
+## Release Note : serial management in separate file
 
 from flask import Flask, render_template, request
 import serial
 import serial.tools.list_ports
-import requests
-import datetime
 import schedule
 import threading
 import time
+import serial_handler
 
 app = Flask(__name__)
 
@@ -22,12 +21,12 @@ def get_serial_ports():
 # Replace with the default serial port
 default_serial_port = '/dev/ttyACM0'
 ser = None
-
-# Attempt to open the default serial port
 try:
     ser = serial.Serial(default_serial_port, 115200)
-except serial.SerialException:
-    pass  # Handle the exception if the default port is not accessible
+    serial_handler.set_serial_object(ser)  # Pass the ser object to serial_handler
+except serial.SerialException as e:
+    print(f"An error occurred while opening the serial port: {e}")
+    ser = None
 
 @app.route('/')
 def index():
@@ -39,32 +38,18 @@ def index():
 
     return render_template('index.html', default_port=default_serial_port, available_ports=available_ports, baud_rates=baud_rates)
 
-
-def read_serial_data(data):
-    if "{hay orange-pi!" in data:
-        if "send time" in data or "update time" in data or "send updated time" in data:
-            update_time()
-        elif "send ip" in data or "update ip" in data or "my ip" in data:
-            ip = requests.get('https://api.ipify.org').text
-            msg = '{hay ttgo-tcall! here is the ip: ' + ip + '.}'
-            print(f"sending : {msg}")
-            send_to_serial_port(msg)
-        else:
-            print(f"unknown keywords in command: {data}")
-
 @app.route('/read_serial')
 def read_serial():
     if ser:
         try:
             data = ser.readline().decode('utf-8')
-            read_serial_data(data)
+            serial_handler.read_serial(data)
 
             return {'data': data.strip()}  # Strip whitespace from the data
         except UnicodeDecodeError:
             return {'error': 'Error decoding serial data'}
     else:
         return {'error': 'Serial port not accessible'}
-
 
 @app.route('/send_serial', methods=['POST'])
 def send_serial():
@@ -73,34 +58,7 @@ def send_serial():
     return {'status': 'success'}
 
 
-def fetch_current_time_online():
-    try:
-        response = requests.get('http://worldtimeapi.org/api/timezone/Asia/Karachi')
-        data = response.json()
-        current_time = datetime.datetime.fromisoformat(data['datetime'])
-        formatted_time = current_time.strftime("%y/%m/%d,%H:%M:%S")
-        return f"py_time:{formatted_time}+20"
-    except requests.exceptions.RequestException as e:
-        print(f"An error occurred: {e}")
-        return None
-
-def send_to_serial_port(serial_data):
-    try:
-        # ser = serial.Serial('/dev/ttyACM0', 115200)
-        ser.write(serial_data.encode())
-        # ser.close()
-    except serial.SerialException as e:
-        print(f"An error occurred while sending data to serial port: {e}")
-
-def update_time():
-    current_time = fetch_current_time_online()
-    if current_time:
-        print(f"Current time in Karachi: {current_time}")
-        send_to_serial_port(current_time)
-    else:
-        print("Failed to fetch current time.")
-
-schedule.every(2).minutes.do(update_time)
+schedule.every(2).minutes.do(serial_handler.update_time)
 
 def update_schedule():
     while True:
